requirements.txt
### RUN sudo pip3 install -r requirements.txt

keyboard
zaber.serial
numpy
pandas
h5py
scipy
opencv-python
imutils
matplotlib
